WorkInProgress: Add midonet-repository charm
This charm will install the necessary midonet repositories.  There
should be another charm that installs MEM and that if they are related
they set up the proper priorities.

The current charm could probably be improved by only installing the
repos when a relation joins that needs a specific repository. For
example, it would only install the neutron plugin repo if the nova
compute relation joins.

WIP!!! Needs testing and refining.

Change-Id: Ieec4de81dae57ee1c335e02fd6e07219dd2f8faf
Signed-off-by: Antoni Segura Puimedon <toni@midokura.com>